malloc error message

+++ srcs/print/ft_putnbr_fd.c
@@ -6,15 +6,15 @@
 /*   By: mamaurai <mamaurai@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/05 11:44:51 by mamaurai          #+#    #+#             */
-/*   Updated: 2021/12/05 11:46:59 by mamaurai         ###   ########.fr       */
+/*   Updated: 2022/02/16 13:40:38 by mamaurai         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	__putnbr(int n, int fd)
+void	__putnbr(int64_t n, int fd)
 {
-	unsigned int	nb;
+	uint64_t	nb;
 
 	if (n < 0)
 		nb = -n;
